feat(tabs): 将组件属性的处理逻辑放到useProperties中
